Откат назад - список перевозчиков не приходит по АПИ
