+++ TravelSchedule/MVVM/ViewModels/ок/CitySelectionViewModel.swift
