+++ TravelSchedule/MVVM/Views/Components/DirectionCardView.swift
@@ -71,5 +71,3 @@ struct DirectionCardView: View {
         .buttonStyle(PlainButtonStyle())
     }
 }
-
-
