+++ TravelSchedule/Resources/OpenAPI /openapi.yaml
@@ -4,7 +4,7 @@ info:
   description: API Яндекс.Расписаний
   version: 3.0.0
 servers:
-- url: https://api.rasp.yandex.net
+- url: https://api.rasp.yandex-net.ru
 tags:
 - name: Список ближайших станций
   description: Список ближайших станций
